+++ kaskade/widgets/topic_list.py
@@ -6,15 +6,15 @@ from rich.text import Text
 from textual import events
 from textual.keys import Keys
 from textual.reactive import Reactive
-from textual.widget import Widget
 
 from kaskade import styles
 from kaskade.kafka.models import Topic
 from kaskade.renderables.scrollable_list import ScrollableList
+from kaskade.widgets.tui_widget import TuiWidget
 
 
-class TopicList(Widget):
-    has_focus: Reactive = Reactive(False)
+class TopicList(TuiWidget):
+    has_focus: Reactive[bool] = Reactive(False)
     scrollable_list: Optional[ScrollableList[Topic]] = None
 
     def max_renderables_len(self) -> int:
@@ -26,9 +26,9 @@ class TopicList(Widget):
             "Topics not found", vertical="middle"
         )
 
-        if self.app.topics:
+        if self.tui.topics:
             self.scrollable_list = ScrollableList(
-                self.app.topics,
+                self.tui.topics,
                 max_len=self.max_renderables_len(),
                 selected=self.scrollable_list.selected
                 if self.scrollable_list
@@ -38,7 +38,7 @@ class TopicList(Widget):
 
         title = Text.from_markup(
             "[bold]topics[/] ([blue]total[/] [yellow]{}[/])".format(
-                len(self.app.topics)
+                len(self.tui.topics)
             )
         )
         return Panel(
@@ -54,22 +54,22 @@ class TopicList(Widget):
             return
 
         self.scrollable_list.next()
-        self.app.topic = self.scrollable_list.selected
+        self.tui.topic = self.scrollable_list.selected
 
-        self.app.enable_describer_mode()
+        self.tui.enable_describer_mode()
 
     def previous(self) -> None:
         if self.scrollable_list is None:
             return
 
         self.scrollable_list.previous()
-        self.app.topic = self.scrollable_list.selected
+        self.tui.topic = self.scrollable_list.selected
 
-        self.app.enable_describer_mode()
+        self.tui.enable_describer_mode()
 
     def on_focus(self) -> None:
         self.has_focus = True
-        self.app.focusables.current = self
+        self.tui.focusables.current = self
 
     def on_blur(self) -> None:
         self.has_focus = False
@@ -83,20 +83,20 @@ class TopicList(Widget):
         self.refresh()
 
     async def on_mouse_scroll_up(self) -> None:
-        await self.app.set_focus(self)
+        await self.tui.set_focus(self)
         self.next()
         self.refresh()
 
     async def on_mouse_scroll_down(self) -> None:
-        await self.app.set_focus(self)
+        await self.tui.set_focus(self)
         self.previous()
         self.refresh()
 
     async def on_click(self, event: events.Click) -> None:
         if self.scrollable_list is not None:
             self.scrollable_list.pointer = event.y - 1
-            self.app.topic = self.scrollable_list.selected
+            self.tui.topic = self.scrollable_list.selected
 
-        self.app.enable_describer_mode()
+        self.tui.enable_describer_mode()
 
         self.refresh()