+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2022-07-10
+### Changed
+- Update help panel
+- Upgrade dependencies
+- Enabling python 3.10
+- Added --configs and --yml commands
+
 ## [0.3.1] - 2022-06-10
 ### Changed
 - Dependencies upgrated
@@ -36,7 +43,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Testing release
 - Showing the messages count and lag
 
-[Unreleased]: https://github.com/sauljabin/kaskade/compare/v0.3.0...HEAD
+[Unreleased]: https://github.com/sauljabin/kaskade/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/sauljabin/kaskade/compare/v0.3.1...v0.4.0
+[0.3.1]: https://github.com/sauljabin/kaskade/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/sauljabin/kaskade/compare/v0.2.9...v0.3.0
 [0.2.9]: https://github.com/sauljabin/kaskade/compare/0.2.8...v0.2.9
 [0.2.8]: https://github.com/sauljabin/kaskade/compare/0.2.7...0.2.8